e2e: fix test for fetch id
Release note: None
